# api/docs/swagger.yaml
definitions:
  models.Err:
    properties:
      error:
        type: string
    type: object
  models.NewWallet:
    properties:
      user_id:
        type: string
      wallet_id:
        type: string
    type: object
  models.User:
    properties:
      email:
        type: string
      first_name:
        type: string
      id:
        type: string
      last_name:
        type: string
      phone_number:
        type: string
    type: object
  models.Wallet:
    properties:
      balance:
        type: number
      id:
        type: string
    type: object
  models.WalletFill:
    properties:
      amount:
        type: number
      id:
        type: string
      user_id:
        type: string
    type: object
info:
  contact: {}
paths:
  /check/wallet/exist:
    post:
      consumes:
      - application/json
      description: Вы можете узнать что есть или нет кашелек
      parameters:
      - description: wallet check existance
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.Wallet'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Wallet'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Err'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Err'
      security:
      - Digest: []
      summary: Check Wallet Existanse
  /user/new:
    post:
      consumes:
      - application/json
      description: здесь вы можете создать нового пользователя
      parameters:
      - description: новый пользователь
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Err'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Err'
      summary: New User
  /wallet/balance:
    post:
      consumes:
      - application/json
      description: Вы можете узнать. Ваше чоте сколько осталось балансе
      parameters:
      - description: wallet check balance
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.Wallet'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Wallet'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Err'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Err'
      security:
      - Digest: []
      summary: Wallet Check Balance
  /wallet/fill:
    post:
      consumes:
      - application/json
      description: По этой конечной точке вы можете заполнить или пополнить свой кошелек
      parameters:
      - description: fill wallet
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.WalletFill'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Wallet'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Err'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Err'
      security:
      - Digest: []
      summary: Wallet Fill
  /wallet/new:
    post:
      consumes:
      - application/json
      description: здесь вы можете создать нового кошелек
      parameters:
      - description: новый кошелек
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.NewWallet'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Wallet'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Err'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Err'
      summary: New Wallet
swagger: "2.0"